Corey   game state game mode (#67)
* Fix Nickname error

* test

* Expand IgameMode, implement in GameMode_Standard

* Fix runtime list of Players, add callbacks for player Enter and Player Leave room

* Modify Player Prefab Model

* Created MultiplayerData structs

* Update PhotonServerSettings.asset

* Test button 'g' for shooting opponent

* Can now find another player during test

* RPC Messaging

* Update RPC OnDamageRequest

* Check if TargetView is me, then don't call the RPC

* RPC callback working!

Can now use RPC callbacks to trigger a function in a single player within a multiplayer scene.

* Failing to Synchronize Timer

* GameTimer decremented If MasterClient

* Fixed Game Timer Synchronization

Every second a player checks if they are the master client. If so, they decrement the Synchronous Timer and broadcast it's value to all other players.

* Comments and Code Cleanup

minor, remove debug logs

More commenting

* Update from main, resolve merge conflicts, setup mic

* Game Ended RPC, JSON data transfer

Created a serializable struct for game score.
Created an RPC callback to announce that the game has ended and what the score is.

* Integrated Shooting with Damage Player

Bullet can detect the ID and Body part of who they shot.

Struct created for damageInfo

The correct player is informed that they have been shot

* Prepping for game score integration

* Failed to integrate HP

* Shooting works for the target

The solution was to get the PV from the raycast and use their PV to RPC and target only that player. They can then inform others that they have been shot.

* Synchronous death works

The health component tells all players that player X has died, the MasterClient will count score, and the gameMode handles respawn.

* Working Timer UI and Game End

* Create Orange

* Team colours working, added 3DModel manager

* Debug Test Print

* remove test code

* Team scores counting correctly

* Score synchronizes at game end

* Started post-match summary scene

* Post match summary working

* More Comments

Comments

* Score updates at every score event, and not on leave

* NumPlayers resets each game

* Leave room before disconnecting

* Remove PhotonView from menu

* Assign TeamNumber to damage

* Get playerID and TeamID from entity on bullet

* Move spawn to up in the air

* Rename method

* Comments

* More postgame info

* Fix merge errors from others scripts

* Scoreboard text fix

* Bug Hunting, error assertions

* Tack on respawn to Player Health

* Maybe fixed score, defo fixed double death

* Minor fix